pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "threshold-engine"
version = "0.1.0"
description = "Session-aware dynamic alerting thresholds for Prometheus/Mimir"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "numpy>=1.26",
    "pandas>=2.2",
    "pydantic>=2.5",
    "pytz>=2024.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
threshold-engine = "threshold_engine.__main__:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
